refactor:(UserController) moving to api folder and adding updated nasmespace
